arrow_array behaviour (arrow v0.1.0)

Provides a record, a behaviour, and functions to work with Apache Arrow Arrays.

This module as mentioned, provides conveniences for working with Apache Arrow Arrays of different Layouts. Firstly, it provides a record to represent all the Layout. Secondly, it provides a behaviour for different layouts to adhere to common functionality. Lastly, it provides functions to work with Arrays in a common manner.

The Structure of an Array

arrow's implementation of an Array has the following fields in its record definition, each corresponding to a field in the Arrow Columnar Format specification:

Field	Type	Representing

layout	atom/0	Layout

type	arrow_type:arrow_type/0	Logical Type

len	pos_integer/0	Length

element_len	pos_integer/0	Length of each element

null_count	non_neg_integer/0	Null Count

validity_bitmap	arrow_buffer:buffer/0	Validity Bitmap

offsets	arrow_buffer:buffer/0	Offsets

data	arrow_buffer:buffer/0	Value Buffer

Certain fields are not required for certain layouts. For example, for the Fixed-Sized Primitive Layout, the offsets field is not required, in which case it is assigned as undefined. Similarly, the validity_bitmap is not required if there are no null values, in which case it is also assigned as undefined.

The Behaviour of an Array

As of right now, a layout needs to implement the from_erlang/2 callback. This is then used in the from_erlang/3 function to create new arrays.

Functions for working with Arrays

As of right now, only functions to access the various fields, to create new arrays, and to serialize arrays to arrow exist.

Types

array()

-type array() ::
          #array{layout :: arrow_array:layout(),
                 type :: arrow_type:arrow_longhand_type() | arrow_type:arrow_nested_type(),
                 len :: pos_integer(),
                 element_len :: pos_integer() | undefined,
                 null_count :: non_neg_integer(),
                 validity_bitmap :: arrow_buffer:buffer() | undefined,
                 offsets :: arrow_buffer:buffer() | undefined,
                 data :: arrow_buffer:buffer() | #array{}}.

Represents an Array.

layout()

-type layout() :: fixed_primitive | variable_binary | fixed_list | variable_list.

Represents the Layout of an Array.

Callbacks

from_erlang(Value, Opts)

-callback from_erlang(Value :: [arrow_type:native_type()], Opts :: map()) -> Array :: array().

Creates a new array of a certain layout, given its value and options from its erlang representation.

Functions

data(Array)

-spec data(Array :: array()) -> Data :: arrow_buffer:buffer() | array() | undefined.

Returns the data of an array.

element_len(Array)

-spec element_len(Array :: array()) -> Length :: pos_integer() | undefined.

Returns the length of an element in an array.

from_erlang(Layout, Value, Opts)

-spec from_erlang(Layout :: layout(),
                  Value :: [arrow_type:native_type()],
                  Opts :: map() | arrow_type:arrow_type()) ->
                     Array :: array().

A common way to create a new array, given its layout, value, and options. from its erlang representation.

layout(Array)

-spec layout(Array :: array()) -> Layout :: layout().

Returns the layout of an array.

len(Array)

-spec len(Array :: array()) -> Length :: pos_integer().

Returns the length of an array.

null_count(Array)

-spec null_count(Array :: array()) -> NullCount :: non_neg_integer().

Returns the null count of an array.

offsets(Array)

-spec offsets(Array :: array()) -> Offsets :: arrow_buffer:buffer() | undefined.

Returns the offsets of an array.

to_arrow(Array)

-spec to_arrow(Array :: array()) -> Arrow :: binary().

Serializes an array into the Arrow binary form.

Serializes the buffers of an Array and concatenates them in the following order:

validity

offsets

data

In case an array doesn't have any of the following buffers, it is omitted. (e.g. validity in arrays with a null count of 0, offsets in fixed primitive arrays). In the case of a nested array, data will be serialized form of nested array.

Do note that this is just binary form that includes the buffers in an Array, and not IPC.

type(Array)

-spec type(Array :: array()) -> Type :: arrow_type:arrow_type().

Returns the type of an array.

validity_bitmap(Array)

-spec validity_bitmap(Array :: array()) -> ValidityBitmap :: arrow_buffer:buffer() | undefined.

Returns the validity bitmap of an array.