arrow_buffer (arrow v0.1.0)

Buffer implementation for arrow.

This module adds support for buffers, or Contiguous Memory Regions. There are multiple things to know about buffers:

Each value it stores is called an element or a slot.

Each slot's length (in bytes) is a positive integer. As a result when we say that a slot has a length of 1, we mean that each slot has a length of 1 byte.

The buffer's length in the metadata refers to the unpadded binary's size in bytes.

All buffers have a size that is a multiple of 64. If their data's length is not a multiple of 64, it must be padded (in this implementation, by zeros).

Null values are represented in this implementation by zeros.

In this implementation buffers can be initialized from raw bytes as data apart datatypes supported by Arrow. This is so that the Validity Bitmap Buffer can be initialized.

Types

buffer()

-type buffer() ::
          #buffer{type :: arrow_type:arrow_longhand_type(),
                  length :: pos_integer(),
                  data :: [arrow_type:native_type()] | binary()}.

Represents an Arrow Buffer.

Functions

from_erlang(Value, Type)

-spec from_erlang(Value :: [arrow_type:native_type()] | binary(),
                  Type :: arrow_type:arrow_longhand_type()) ->
                     Buffer :: arrow_buffer:buffer().

Creates a new buffer from a list of Erlang values or binaries, given its type.

from_erlang(Data, Type, DataLen)

-spec from_erlang(Data :: [arrow_type:native_type()] | binary(),
                  Type :: arrow_type:arrow_longhand_type(),
                  DataLen :: pos_integer() | undefined) ->
                     Buffer :: arrow_buffer:buffer().

Creates a new buffer from a list of Erlang values or binaries, given its type and length

size(Buffer)

-spec size(Buffer :: arrow_buffer:buffer()) -> pos_integer().

Returns the size of the buffer inclusive of padding in bytes.

to_arrow(Buffer)

-spec to_arrow(Buffer :: arrow_buffer:buffer()) -> binary().

Returns an Arrow buffer binary given a buffer.

to_erlang(Buffer)

-spec to_erlang(Buffer :: arrow_buffer:buffer()) -> [arrow_type:native_type()].

Returns a list of Erlang values or binaries from a buffer.