arrow_ipc_field (arrow v0.1.0)

Provides a record and functions to deal with individual fields in a Schema

A Field represents a single column in a table (which is represented by a Schema. This module provides a record and a function to manage all the metadata required to represent a column. Metadata such as:

name: The Name of the column. This can either be a string or null.

nullable: A boolean representing whether a column can have null values. Generally true.

type: The Type of the column, and is of type arrow_ipc_type:ipc_type/0

dictionary: A boolean representing if the column is dictionary encoded

children: The child fields of a nested datatype

custom_metadata: A list of custom metadata in key-value format

Currently, dictionary encoding and custom metadata are not supported, but they have been added for forwards comapatibility.

Types

field()

-type field() ::
          #field{name :: string() | undefined,
                 nullable :: boolean(),
                 type :: arrow_ipc_type:ipc_type(),
                 dictionary :: term(),
                 children :: [#field{}],
                 custom_metadata :: [arrow_ipc_message:key_value()]}.

Represents an IPC Field.

Functions

from_erlang(Type)

-spec from_erlang(Type :: arrow_ipc_type:ipc_type()) -> Field :: field().

Creates a field given the type of the column. Assigns the name as undefined.

from_erlang(Type, Name)

-spec from_erlang(Type :: arrow_ipc_type:ipc_type(), Name :: string() | undefined) -> Field :: field().

Creates a field given the type and name of the column.

from_erlang(Type, Name, Children)

-spec from_erlang(Type :: arrow_ipc_type:ipc_type(),
                  Name :: string() | undefined,
                  Children :: [field()]) ->
                     Field :: field().

Creates a field given the type, name and children of the column.