arrow_ipc_message (arrow v0.1.0)

Provides a record and functions to deal with the Encapsulated Message Format.

A Message is a serialized form of a Schema or a RecordBatch (which may be required to read a serialized array) along with some metadata. This module provides a record and a function to manage all the metadata required to represent a message. Metadata such as:

version: The Apache Arrow Format Version. One of v1..v5. Defaults to v5.

header: The metadata of the Schema or RecordBatch

body_length: The length of the body in bytes

custom_metadata: A list of custom metadata in key-value format

body: The actual body. Can be undefined (in the case of Schema) or a binary (in the case of Record Batch).

Currently, changing the version and custom metadata are not supported, but they have been added for forwards compatibility.

This module also provides the to_ipc/1 function which serializes the message into the Encapsulated Message Format. However, this function gives incomplete output with invalid metadata, because of an unsatisfied dependency on flatbuffers, which is required for serializing the metadata.

Types

key_value()

-type key_value() :: #{key => string(), value => string()}.

Key-Value structure for custom metadata. See the definition for more info:

message()

-type message() ::
          #message{version :: arrow_ipc_message:metadata_version(),
                   header ::
                       #schema{endianness :: arrow_ipc_schema:endianness(),
                               fields :: term(),
                               custom_metadata :: [arrow_ipc_message:key_value()],
                               features :: [arrow_ipc_schema:feature()]} |
                       #record_batch{length :: pos_integer(),
                                     nodes :: [arrow_ipc_record_batch:field_node()],
                                     buffers :: [arrow_ipc_record_batch:buffer()],
                                     compression :: undefined | lz4_frame | zstd},
                   body_length :: non_neg_integer(),
                   custom_metadata :: [arrow_ipc_message:key_value()],
                   body :: binary() | undefined}.

Represents a Message.

metadata_version()

-type metadata_version() :: v1 | v2 | v3 | v4 | v5.

The Arrow version. See the definition for more info:

Functions

body_from_erlang(Columns)

-spec body_from_erlang(Columns :: [arrow_array:array()]) -> Body :: binary().

Returns the body of message from a list of arrays.

Shorthand for:

<<<<(arrow_array:to_arrow(Array))/binary>> || Array <- Columns>>

from_erlang(Header)

-spec from_erlang(Header :: arrow_ipc_schema:schema() | arrow_ipc_record_batch:record_batch()) ->
                     Message :: message().

Creates a message given a data header.

from_erlang(Header, Body)

-spec from_erlang(Header :: arrow_ipc_schema:schema() | arrow_ipc_record_batch:record_batch(),
                  Body :: binary()) ->
                     Message :: message().

Creates a message given a data header and a body.

metadata_len(EMF)

-spec metadata_len(EMF :: binary()) -> MetadataLen :: non_neg_integer().

Returns the Metadata Length of an EMF.

to_ipc(Message)

-spec to_ipc(Message :: message()) -> EMF :: binary().

Serializes a message into the Encapsulated Message Format.

to_stream(Messages)

-spec to_stream(Messages :: [message()] | [binary()]) -> Stream :: binary().

Serializes a list of messages or EMFs into a Stream.