arrow_ipc_schema (arrow v0.1.0)

Provides a record and functions to deal with Schemas.

A Schema represents a table, or a list of arrays of equal length. This module provides a record and a function to manage all the metadata required to represent a schema. Metadata such as:

endianness: The Endianness of the table. One of little or big. Defaults to little.

fields: The list of fields in a table.

type: The Layout of the column

custom_metadata: A list of custom metadata in key-value format

features: Any features used by the table which may not be present in other implementations of Arrow.

Currently, big endianness, custom metadata and features are not supported, but they have been added for forwards compatibility.

You can find Schemas in the Arrow spec here.

Types

endianness()

-type endianness() :: little | big.

Endianness of the data. Either little or big.

feature()

-type feature() :: unused | dictionary_replacement | compressed_body.

Features used in the data which may not be present in other implementations. See the definition.

schema()

-type schema() ::
          #schema{endianness :: arrow_ipc_schema:endianness(),
                  fields :: term(),
                  custom_metadata :: [arrow_ipc_message:key_value()],
                  features :: [arrow_ipc_schema:feature()]}.

Represents a schema

Functions

from_erlang(Fields)

-spec from_erlang(Fields :: [arrow_ipc_field:field()]) -> Schema :: schema().

Creates a Schema given an ordered list of fields.